feat: clean up default exposed API
There were a multitude of items being exposed by default through the main pagckage file. Aiming to clarify the main utilities being exposed by this package, now only `is`, `isOptions` and `DataType` are being exposed through the main file, which makes the package intent much clearer, in my opinion.

It’s worth noting, however, that the remainder of the functions and interfaces are still accessible if needed.