perf: create a sister const enum for DataType for internal use
This change should improve performace due to removing the overhead of object access in favor of primitive use in comparisons. Additionally, becase DATATYPE is compliled into just numbers, there will be a size reduction.

DATATYPE will not be exported to JS, only DataType.

A test has been added to ensure that both DATATYPE and DataType stay in sync.